+++ src/main/java/domain/result/RoundBetInfo.java
@@ -2,9 +2,9 @@ package domain.result;
 
 import domain.player.User;
 
-public record RoundBetInfo(int round, User user, int betAmount) {
+public record RoundBetInfo(int round, User user, double betAmount) {
 
-    public UserProfit toUserProfit(GameResult gameResult, int profit) {
+    public UserProfit toUserProfit(GameResult gameResult, double profit) {
         return new UserProfit(this, gameResult,profit);
     }
 }
